feat(237-01): extract shared search filter foundation
- move SearchFilters, FacetCounts, OGC datetime parsing, text filter construction, and common filters into service_filters.py

- keep app.modules.catalog.search.service as the stable facade for existing imports
